requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
Flask-Swagger-UI
psycopg2-binary
